requirements.txt
# SUBFRACTURE LangGraph Platform Dependencies

# Core LangGraph & LangSmith (Platform optimized)
langgraph>=0.2.0
langsmith>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0

# Platform CLI
langgraph-cli>=0.1.0

# State Management & Validation
pydantic>=2.0.0
typing-extensions>=4.5.0

# Async & HTTP
asyncio
aiohttp>=3.9.0
httpx>=0.27.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration & Environment
python-dotenv>=1.0.0
dynaconf>=3.2.0

# Logging & Monitoring  
structlog>=23.0.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# System monitoring and performance
psutil>=5.9.0

# Advanced memory management and knowledge graphs
cognee>=0.1.0